move date stackers into their own submodule
