+++ rubin_sim/maf/stackers/__init__.py
@@ -1,5 +1,6 @@
 from .base_stacker import *
 from .coord_stackers import *
+from .date_stackers import *
 from .general_stackers import *
 from .get_col_info import *
 from .label_stackers import *
